refactor(map): split enrichment out of EnrichedLayerFeatures
EnrichedLayerFeatures.tsx had grown to 673 lines holding four unrelated
jobs, with the config interfaces declared at line 437 -- well after the
builders that consume them. It is now a 97-line React component over a
`enrichment/` module per concern:

  types.ts         the config shapes and the synthetic-feature alias
  arrow.ts         sprites, rotation convention, shared arrow primitives
  slideArrow.ts    the Rutschgebiet arrow and its tuning
  flowArrow.ts     the flood-zone arrow, ring helpers and its tuning
  registry.ts      which line and zone types get which indicator
  enrichFeature.ts the dispatch over a feature's geometry

The public entry point is unchanged -- EnrichedLayerFeatures still
re-exports enrichFeature, EnrichLineStringMap and EnrichPolygonMap -- so
no import site moved.

Simplifications taken along the way, all behaviour-preserving:

- SlideArrowConfig and FlowArrowConfig were structurally identical, so
  TypeScript treated them as the same type: the split bought commentary
  and no safety at all. Merged into ArrowConfig, with what `lengthRatio`
  is measured against documented at each registry entry, since that is
  the part that genuinely differs.
- The length clamp and the self-crossing probe were duplicated between
  the two builders. Both are now `clampArrowLength` and
  `spanBeforeCrossing` in arrow.ts, which also stops the two
  self-crossing ratios drifting apart unnoticed.
- The start and end cap blocks were near-copies. Both measure their
  bearing *inwards*, which is exactly why they add the rotation offset
  where the arrows subtract it, so `capFeature` now states that once.
- A shared KM units constant replaces ~15 inline `{ units: "kilometers" }`.
- damageExtent used a bare 90 where CHEVRON_BEARING_OFFSET was meant.
- EnrichedSymbolSource built a collection then mutated its features; it
  now builds the whole thing inside the memo.
- The `selectedFeature` prop has been dead since enrichment stopped
  excluding the selected feature, so it is gone from the interface and
  its two call sites.

201 tests pass unchanged, and oxlint reports no warnings in the tree.
